Fix brightness handling and streamline logging
