+++ is_matrix_forge/led_matrix/display/effects/breather.py
@@ -1,6 +1,5 @@
 import threading
 import time
-import is_matrix_forge.led_matrix.display.helpers
 from is_matrix_forge.common.decorators import freeze_setter
 from is_matrix_forge.log_engine import ROOT_LOGGER, Loggable
 
@@ -51,22 +50,20 @@ class Breather(Loggable):
         self._step           = None
         self._fps            = None
 
-        self.controller      = controller
+        self.controller = controller
         log = self.class_logger
-        log.debug(f'Controller: {controller}')
-
         self.__initial_brightness = self.controller.brightness
-        log.debug(f'Initial brightness: {self.__initial_brightness}')
-
-        self.min_brightness  = min_brightness
-        self.max_brightness  = max_brightness
-        log.debug(f'Min brightness: {self.min_brightness} | Max brightness: {self.max_brightness}')
-
-        self.step            = step
-        log.debug(f'Step: {self.step}%')
-
-        self.fps             = breathe_fps
-        log.debug(f'FPS: {self.fps}')
+        self.min_brightness = min_brightness
+        self.max_brightness = max_brightness
+        self.step = step
+        self.fps = breathe_fps
+        log.debug(
+            "Breather initialized (min=%s max=%s step=%s fps=%s)",
+            self.min_brightness,
+            self.max_brightness,
+            self.step,
+            self.fps,
+        )
 
         self._breathing = False
         self._thread    = None
@@ -80,13 +77,9 @@ class Breather(Loggable):
         if not isinstance(new, bool):
             raise TypeError(f'"breathing" must be of type `bool`, not {type(new)}')
 
-        self.method_logger.debug(f'Received new value for "breathing": {new}')
-
         if self.breathing and not new:
-            self.method_logger.debug('Stopping breathing')
             self.stop()
         elif not self.breathing and new:
-            self.method_logger.debug('Starting breathing')
             self.start()
 
         self._breathing = new
@@ -105,8 +98,6 @@ class Breather(Loggable):
         from is_matrix_forge.led_matrix.controller.controller import LEDMatrixController
         if not isinstance(new, LEDMatrixController):
             raise TypeError(f'controller must be LEDMatrixController, not {type(new)}')
-
-        self.method_logger.debug(f'New controller: {new}')
 
         self._controller = new
 
@@ -135,8 +126,6 @@ class Breather(Loggable):
         if self._max_brightness is not None and new > self._max_brightness:
             raise ValueError('"min_brightness" must be <= max_brightness')
 
-        self.method_logger.debug(f'New min brightness: {new}')
-
         self._min_brightness = new
 
     @property
@@ -155,7 +144,6 @@ class Breather(Loggable):
         if self._min_brightness is not None and new < self._min_brightness:
             raise ValueError('"max_brightness" must be >= min_brightness')
 
-        self.method_logger.debug(f'New max brightness: {new}')
         self._max_brightness = new
 
     @property
@@ -172,8 +160,6 @@ class Breather(Loggable):
         if new <= 0 or new > 100:
             raise ValueError('"step" must be > 0 and <= 100')
 
-        self.method_logger.debug(f'New step: {new}')
-
         self._step = new
 
     @property
@@ -189,8 +175,6 @@ class Breather(Loggable):
             raise TypeError('"fps" must be a number')
         if new <= 0:
             raise ValueError('"fps" must be > 0')
-
-        self.method_logger.debug(f'New FPS: {new}')
 
         self._fps = float(new)
